/ui: split utils
Closes #437
